pyproject.toml
[project]
name = "trip-sage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Krak3nDev", email = "bihichnazar@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "colorlog>=6.9.0",
    "dishka>=1.6.0",
    "fastapi>=0.115.14",
    "load-dotenv>=0.1.0",
    "openai>=1.93.0",
    "sqlalchemy>=2.0.41",
    "tenacity>=9.1.2",
    "uvicorn>=0.35.0",
]


[project.scripts]
trip-sage = "trip_sage:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.12.2",
]



[tool.ruff]
line-length = 79
indent-width = 4
target-version = "py312"
[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "E501", "I"]
ignore = []
fixable = ["ALL"]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"


exclude = [
  "tests/",
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]
